@@ -1,9 +1,11 @@
 package com.subcrowd.app.Chat;
 
-import android.support.v7.widget.RecyclerView;
+
 import android.view.View;
 import android.widget.LinearLayout;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.subcrowd.app.R;
 
