@@ -1,10 +1,12 @@
 package com.subcrowd.app;
 
 import android.content.Intent;
-import android.support.v7.app.AppCompatActivity;
+
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 public class ChooseLoginRegistrationActivity extends AppCompatActivity {
 
